// src/Types/StudentsListItemInterface.ts
export interface StudentsListItemInterface {
  id: string;
  courseCompletion: number;
  courseEngagement: number;
  projectDegree: number;
  teamProjectDegree: number;
  firstName: string;
  lastName: string;
  targetWorkCity: string;
  expectedSalary: number;
  canTakeApprenticeship: number;
  monthsOfCommercialExp: number;
  workExperience: string;
  typeContract: string;
}